+++ src/webgis4u/ol/util.js
@@ -2,4 +2,5 @@
  * @module webgis4u/ol/util
  */
 
+export * from './util/getMapType';
 export * from './util/getScale';